requirements.txt
pandas
pillow
playwright